refactor: 시트 파싱 로직 공통화 (#31)
